collision fix

+++ src/Math.cpp
@@ -4,5 +4,5 @@
 float to_radians(float degrees) { return degrees / 180.0f * M_PI; }
 
 float random_range(float minv, float maxv) {
-  return minv + ((float)rand() / (RAND_MAX / (maxv - minv)));
+  return minv + (rand() / (RAND_MAX / (maxv - minv)));
 }
